+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/common/Tag.kt
@@ -1,5 +1,8 @@
 package io.bkbn.kompendium.oas.common
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class Tag(
   val name: String,
   val description: String? = null,