+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/AnyOfSchema.kt
@@ -1,3 +1,6 @@
 package io.bkbn.kompendium.oas.schema
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class AnyOfSchema(val anyOf: List<ComponentSchema>, override val description: String? = null) : ComponentSchema