+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/ComponentSchema.kt
@@ -1,5 +1,10 @@
 package io.bkbn.kompendium.oas.schema
 
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.json.JsonClassDiscriminator
+
+@OptIn(ExperimentalSerializationApi::class)
+@JsonClassDiscriminator("component_type") // todo figure out a way to filter this
 sealed interface ComponentSchema {
   val description: String?
     get() = null