+++ kompendium-playground/src/main/kotlin/io/bkbn/kompendium/playground/ExceptionPlayground.kt
@@ -6,12 +6,8 @@ import io.bkbn.kompendium.core.metadata.ExceptionInfo
 import io.bkbn.kompendium.core.metadata.ResponseInfo
 import io.bkbn.kompendium.core.metadata.method.GetInfo
 import io.bkbn.kompendium.core.routes.redoc
-import io.bkbn.kompendium.oas.OpenApiSpec
-import io.bkbn.kompendium.oas.info.Contact
-import io.bkbn.kompendium.oas.info.Info
-import io.bkbn.kompendium.oas.info.License
-import io.bkbn.kompendium.oas.server.Server
 import io.bkbn.kompendium.playground.ExceptionPlaygroundToC.simpleGetExample
+import io.bkbn.kompendium.playground.util.Util
 import io.ktor.application.Application
 import io.ktor.application.call
 import io.ktor.application.install
@@ -25,7 +21,6 @@ import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
 import kotlin.reflect.typeOf
 import kotlinx.serialization.Serializable
-import java.net.URI
 import java.time.LocalDateTime
 
 // Application Entrypoint
@@ -41,11 +36,11 @@ fun main() {
 private fun Application.mainModule() {
   // Installs Simple JSON Content Negotiation
   install(ContentNegotiation) {
-    json()
+    json(json = Util.kotlinxConfig)
   }
   // Installs the Kompendium Plugin and sets up baseline server metadata
   install(Kompendium) {
-    spec = ExceptionMetadata.spec
+    spec = Util.baseSpec
   }
   install(StatusPages) {
     exception<ExceptionModels.BadUserException> {
@@ -68,36 +63,6 @@ private fun Application.mainModule() {
       }
     }
   }
-}
-
-object ExceptionMetadata {
-  val spec = OpenApiSpec(
-    info = Info(
-      title = "Simple Demo API with notarized exceptions",
-      version = "1.33.7",
-      description = "Wow isn't this cool?",
-      termsOfService = URI("https://example.com"),
-      contact = Contact(
-        name = "Homer Simpson",
-        email = "chunkylover53@aol.com",
-        url = URI("https://gph.is/1NPUDiM")
-      ),
-      license = License(
-        name = "MIT",
-        url = URI("https://github.com/bkbnio/kompendium/blob/main/LICENSE")
-      )
-    ),
-    servers = mutableListOf(
-      Server(
-        url = URI("https://myawesomeapi.com"),
-        description = "Production instance of my API"
-      ),
-      Server(
-        url = URI("https://staging.myawesomeapi.com"),
-        description = "Where the fun stuff happens"
-      )
-    )
-  )
 }
 
 // This is a table of contents to hold all the metadata for our various API endpoints